split metrics_test into module tests
